feat: replace DevMenu with AdminMenu and update localization files
