+++ README.md
@@ -520,4 +520,4 @@ erDiagram
 
 ## Lizenz
 
-GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
+GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007k
