+++ src/client/locales/ar.json
@@ -1,9 +1,11 @@
 {
+  "reports.saved.success.toast": "تم حفظ التقرير بنجاح",
+  "reports.saved.failed.toast": "فشل في حفظ التقرير",
   "nav.home": "الصفحة الرئيسية",
   "nav.new_report": "تقرير جديد",
   "nav.admin_dashboard": "لوحة الإدارة",
   "nav.admin_reports": "تقارير المسؤول",
-  "nav.dev_menu": "قائمة المطورين",
+  "nav.admin_menu": "قائمة المطورين",
   "nav.welcome": "مرحباً، {username}",
   "nav.profile": "الملف الشخصي",
   "nav.logout": "تسجيل الخروج",
@@ -127,15 +129,15 @@
   "report.edit": "تعديل",
   "report.delete": "حذف",
   "edit.save": "حفظ التغييرات",
-  "dev.title": "قائمة المطورين",
-  "dev.user_management": "إدارة المستخدمين",
-  "dev.create_user": "إنشاء مستخدم جديد",
-  "dev.existing_users": "المستخدمون الحاليون",
-  "dev.no_users": "لم يتم العثور على مستخدمين.",
-  "dev.action_save": "حفظ",
-  "dev.action_cancel": "إلغاء",
-  "dev.action_edit": "تعديل",
-  "dev.action_change_password": "تغيير كلمة المرور",
-  "dev.action_delete": "حذف",
-  "dev.action_save_password": "حفظ كلمة المرور"
+  "admin_menu.title": "قائمة المطورين",
+  "admin_menu.user_management": "إدارة المستخدمين",
+  "admin_menu.create_user": "إنشاء مستخدم جديد",
+  "admin_menu.existing_users": "المستخدمون الحاليون",
+  "admin_menu.no_users": "لم يتم العثور على مستخدمين.",
+  "admin_menu.action_save": "حفظ",
+  "admin_menu.action_cancel": "إلغاء",
+  "admin_menu.action_edit": "تعديل",
+  "admin_menu.action_change_password": "تغيير كلمة المرور",
+  "admin_menu.action_delete": "حذف",
+  "admin_menu.action_save_password": "حفظ كلمة المرور"
 }
