+++ src/client/locales/nl.json
@@ -1,9 +1,11 @@
 {
+  "reports.saved.success.toast": "Rapport succesvol opgeslagen",
+  "reports.saved.failed.toast": "Fout bij het opslaan van het rapport",
   "nav.home": "Startpagina",
   "nav.new_report": "Nieuw rapport",
   "nav.admin_dashboard": "Admin-dashboard",
   "nav.admin_reports": "Admin-rapporten",
-  "nav.dev_menu": "Dev-menu",
+  "nav.admin_menu": "Dev-menu",
   "nav.welcome": "Welkom, {username}",
   "nav.profile": "Profiel",
   "nav.logout": "Uitloggen",
@@ -127,15 +129,15 @@
   "report.edit": "Bewerken",
   "report.delete": "Verwijderen",
   "edit.save": "Wijzigingen opslaan",
-  "dev.title": "Dev Menu",
-  "dev.user_management": "Gebruikersbeheer",
-  "dev.create_user": "Nieuwe gebruiker maken",
-  "dev.existing_users": "Bestaande gebruikers",
-  "dev.no_users": "Geen gebruikers gevonden.",
-  "dev.action_save": "Opslaan",
-  "dev.action_cancel": "Annuleren",
-  "dev.action_edit": "Bewerken",
-  "dev.action_change_password": "Wachtwoord wijzigen",
-  "dev.action_delete": "Verwijderen",
-  "dev.action_save_password": "Wachtwoord opslaan"
+  "admin_menu.title": "Dev Menu",
+  "admin_menu.user_management": "Gebruikersbeheer",
+  "admin_menu.create_user": "Nieuwe gebruiker maken",
+  "admin_menu.existing_users": "Bestaande gebruikers",
+  "admin_menu.no_users": "Geen gebruikers gevonden.",
+  "admin_menu.action_save": "Opslaan",
+  "admin_menu.action_cancel": "Annuleren",
+  "admin_menu.action_edit": "Bewerken",
+  "admin_menu.action_change_password": "Wachtwoord wijzigen",
+  "admin_menu.action_delete": "Verwijderen",
+  "admin_menu.action_save_password": "Wachtwoord opslaan"
 }
