+++ src/client/locales/uk.json
@@ -1,10 +1,12 @@
 {
+  "reports.saved.success.toast": "Звіт успішно збережено",
+  "reports.saved.failed.toast": "Не вдалося зберегти звіт",
   "//coment//": "Ihave no Idear if the translator work like it was supose to ",
   "nav.home": "Головна",
   "nav.new_report": "Новий звіт",
   "nav.admin_dashboard": "Панель адміністратора",
   "nav.admin_reports": "Звіти адміністратора",
-  "nav.dev_menu": "Меню розробника",
+  "nav.admin_menu": "Меню розробника",
   "nav.welcome": "Вітаємо, {username}",
   "nav.profile": "Профіль",
   "nav.logout": "Вийти",
@@ -128,15 +130,15 @@
   "report.edit": "Редагувати",
   "report.delete": "Видалити",
   "edit.save": "Зберегти зміни",
-  "dev.title": "Меню розробника",
-  "dev.user_management": "Управління користувачами",
-  "dev.create_user": "Створити нового користувача",
-  "dev.existing_users": "Існуючі користувачі",
-  "dev.no_users": "Користувачів не знайдено.",
-  "dev.action_save": "Зберегти",
-  "dev.action_cancel": "Скасувати",
-  "dev.action_edit": "Редагувати",
-  "dev.action_change_password": "Змінити пароль",
-  "dev.action_delete": "Видалити",
-  "dev.action_save_password": "Зберегти пароль"
+  "admin_menu.title": "Меню розробника",
+  "admin_menu.user_management": "Управління користувачами",
+  "admin_menu.create_user": "Створити нового користувача",
+  "admin_menu.existing_users": "Існуючі користувачі",
+  "admin_menu.no_users": "Користувачів не знайдено.",
+  "admin_menu.action_save": "Зберегти",
+  "admin_menu.action_cancel": "Скасувати",
+  "admin_menu.action_edit": "Редагувати",
+  "admin_menu.action_change_password": "Змінити пароль",
+  "admin_menu.action_delete": "Видалити",
+  "admin_menu.action_save_password": "Зберегти пароль"
 }
